+++ packages/apollos-data-connector-bible/src/index.js
@@ -1,5 +1,0 @@
-import resolver from './resolver';
-import dataSource from './data-source';
-
-export { resolver, dataSource };
-export { scriptureSchema as schema } from '@apollosproject/data-schema';