+++ packages/apollos-data-connector-onesignal/src/resolver.js
@@ -1,6 +1,0 @@
-export default {
-  Mutation: {
-    updateUserPushSettings: (root, { input }, { dataSources }) =>
-      dataSources.OneSignal.updatePushSettings(input),
-  },
-};