+++ packages/apollos-data-connector-postgres/src/content-item-categories/migrations/index.js
@@ -1,9 +1,0 @@
-import CreateContentItemsCategories001 from './001_create_content_item_categories';
-import RenameContentItemCategoryTable from './002_rename_content_item_categories_table';
-
-const migrations = [
-  CreateContentItemsCategories001,
-  RenameContentItemCategoryTable,
-];
-
-export default migrations;