+++ packages/apollos-data-connector-postgres/src/content-items/index.js
@@ -1,5 +1,0 @@
-export * as models from './model';
-export dataSource from './dataSource';
-export resolver from './resolver';
-export migrations from './migrations';
-export { contentItemSchema as schema } from '@apollosproject/data-schema';