+++ packages/apollos-data-connector-postgres/src/interactions/migrations/index.js
@@ -1,6 +1,0 @@
-import CreateInteraction001 from './001_create_interaction';
-import UpdatePeopleInteractionForeignKey002 from './002_update_people_interaction_foreign_key';
-
-const migrations = [CreateInteraction001, UpdatePeopleInteractionForeignKey002];
-
-export default migrations;