+++ packages/apollos-data-connector-rock/src/interactions/index.js
@@ -1,7 +1,0 @@
-/* eslint-disable import/prefer-default-export */
-import { interactionsSchema as schema } from '@apollosproject/data-schema';
-
-import dataSource from './data-source';
-import resolver from './resolver';
-
-export { dataSource, schema, resolver };