+++ packages/apollos-ui-auth/src/ProfileDetails/index.js
@@ -1,2 +1,0 @@
-export ProfileDetailsEntry from './ProfileDetailsEntry';
-export ProfileDetailsEntryConnected from './ProfileDetailsEntryConnected';