+++ packages/apollos-ui-auth/src/mutations/authenticate.js
@@ -1,9 +1,0 @@
-import { gql } from '@apollo/client';
-
-export default gql`
-  mutation authenticate($email: String!, $password: String!) {
-    authenticate(identity: $email, password: $password) {
-      token
-    }
-  }
-`;