+++ packages/apollos-ui-onboarding/src/LandingSwiper/LandingSwiper.stories.js
@@ -1,0 +1,7 @@
+import React from 'react';
+import { storiesOf } from '@apollosproject/ui-storybook';
+import LandingSwiper from './index';
+
+storiesOf('Onboarding/LandingSwiper', module).add('full', () => (
+  <LandingSwiper />
+));